Fixed invalid instruction using kick in console
OnPlayerDisconnect is bugged if it takes to much time to return a
value
the server will call the function again whit the same playerid
and
reason if you are lucky.
You can say so what a function is called
twice what's  wrong with it?
Yeah the bad news IS:
- (*)
- the function
is called first time (lets say fnc-A)
- in the middle (not quite in the
middle but near the mid or the end of
the fnc) of the function the
process is stopped
- and now he calls the function for the second time
(fnc-B)
- again in the middle of the procces of fnc-B and continues whit
the
rest of fnc-A
- randomly do again from the mark (*)
- fnc-A end and
then continue whit fnc-B
- and don't forget the fnc is called twcie in
all the gamemodes and
filterscripts (ohh god kill me now)
- and bonus
'Invalid instruction' (i believe there are more then one
type of error)
whit a chance of 80% to be inserted one of the steps above

Lets take a
moment to say tnx Kalcor for not fixing this issue.
PS: To recreate this
bug is a pain in the ass.